opt: rewrote cyno tf
