+++ mona_core/src/target_functions/target_function_name.rs
@@ -80,6 +80,5 @@ pub enum TargetFunctionName {
     KukiShinobuDefault,
     ShikanoinHeizouDefault,
     TighnariDefault,
-    CynoAggravate,
-    CynoMono,
+    CynoDefault,
 }
