+++ mona_core/src/target_functions/target_functions/electro/mod.rs
@@ -8,8 +8,7 @@ pub mod raiden_shogun_default;
 pub mod razor_default;
 pub mod yae_miko_default;
 pub mod kuki_shinobu_default;
-pub mod cyno_aggravate;
-pub mod cyno_mono;
+pub mod cyno_default;
 
 pub use beidou_default::BeidouDefaultTargetFunction;
 pub use fischl_default::FischlDefaultTargetFunction;
@@ -21,5 +20,4 @@ pub use raiden_shogun_default::RaidenShogunDefaultTargetFunction;
 pub use razor_default::RazorDefaultTargetFunction;
 pub use yae_miko_default::YaeMikoDefaultTargetFunction;
 pub use kuki_shinobu_default::KukiShinobuDefaultTargetFunction;
-pub use cyno_aggravate::CynoAggravateTargetFunction;
-pub use cyno_mono::CynoMonoTargetFunction;
+pub use cyno_default::CynoDefaultTargetFunction;
